fix(OI-27): ship ONNX in the wheels + fix the exporter defects the suite was hiding
Ship ONNX (packaging) AND make tests/python/onnx fully green: 10525 passed, 0 failed
(was 83 failed once the suite was actually run end-to-end against onnx 1.21 / ort 1.24).

Packaging:
- tools/build_cleanup_wheel.sh no longer sets MXNET_SETUP_EXCLUDE_ONNX, so the wheel
  bundles the mxnet.onnx / mxnet.contrib.onnx packages.
- setup.py exposes an optional `onnx` extra: `pip install "mxnet[onnx]"` (pure-Python,
  no native rebuild). Pinned onnx<1.22 to the validated range.
- run_macos_wheel_full_test.sh comment updated (ONNX now ships).

Exporter defects (pre-existing; surfaced by running the full suite):
- _np_reshape had no ONNX converter (it is what np.squeeze lowers to). Registered one
  mapping to ONNX Reshape. Fixes 20 squeeze cases.
- Integer sum/prod declared int32 output, but mxnet upcasts integer reductions to int64
  (numpy convention) while ONNX ReduceSum/ReduceProd keep int32 -> graph/node type
  mismatch. Cast integer inputs to int64 and declare int64. Fixes 32 cases.
- mean always declared float32 output instead of preserving the input float dtype, so a
  float16-input mean mismatched its declared float16 graph output. Now compute in fp32
  for precision but cast/declare the true output dtype. Fixes 16 cases.
- tests/python/onnx gather_nd test cast the *indices* to the float data dtype, which
  mxnet's gather_nd rejects (indices must be int32/int64). Use int64 indices. Fixes 3.

Verified on a local build: tests/python/onnx 10525 passed / 0 failed (onnx 1.21 /
ort 1.24); full test_numpy_op.py sweep 11390 passed (nothing broke).

+++ tests/python/onnx/test_operators.py
@@ -774,14 +774,16 @@ def test_onnx_export_npx_reshape_like(tmp_path, dtype):
 
 @pytest.mark.parametrize('dtype', ['int32', 'int64', 'float16', 'float32', 'float64'])
 def test_onnx_export_npx_gather_nd(tmp_path, dtype):
+    # gather_nd indices must be int32/int64 regardless of the data dtype (mxnet
+    # enforces this); the data dtype is what we parametrize over.
     # y[0] == dim(x)
     x1 = mx.np.random.uniform(-100, 100, (4, 5, 6, 7)).astype(dtype)
-    y1 = mx.np.random.randint(-4, 4, (4, 4, 4)).astype(dtype)
+    y1 = mx.np.random.randint(-4, 4, (4, 4, 4)).astype('int64')
     M1 = def_model(mx.npx, 'gather_nd')
     op_export_test('gather_nd1', M1, [x1, y1], tmp_path)
     # y[0] < dim(x)
     x2 = mx.np.random.uniform(-100, 100, (4, 5, 6, 7)).astype(dtype)
-    y2 = mx.np.random.randint(-4, 4, (2,3,4)).astype(dtype)
+    y2 = mx.np.random.randint(-4, 4, (2,3,4)).astype('int64')
     M2 = def_model(mx.npx, 'gather_nd')
     op_export_test('gather_nd2', M2, [x2, y2], tmp_path)
